+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added 
 
  - `lena_speaker` for the LENA its converter
+ - `lena_speaker` aggregated metrics for the LenaMetrics pipeline
+ - Improved AclewMetrics and LenaMetrics performance
  - Improved error handling (dataframes sanity checks)
 
 ### Fixed